requirements.txt
fastapi[all]
uvicorn[standard]
SQLAlchemy
pydantic-settings

fastapi-users[sqlalchemy]
aiosqlite

